Unstable doc build

+++ .travis/deploy_docs.sh
@@ -27,9 +27,15 @@ git clone git@github.com:fabric-composer/${REPO}.git
 git remote set-url origin ${REPO}.git
 
 cd "${DIR}/packages/composer-website/out/${REPO}"
+  
+mkdir -p "${DIR}/packages/composer-website/out/${REPO}/unstable"
 
-rm -rf ${DIR}/packages/composer-website/out/${REPO}/*
-cp -rf ${DIR}/packages/composer-website/jekylldocs/_site/* .
+if [ "${DOCS}" == "full" ]; then
+    rm -rf ${DIR}/packages/composer-website/out/${REPO}/*
+    cp -rf ${DIR}/packages/composer-website/jekylldocs/_site/* .     
+fi
+
+cp -rf ${DIR}/packages/composer-website/jekylldocs/_site/* ./unstable  
 
 git add .
